Use int64 type for timestamp values

+++ timestamp.go
@@ -5,6 +5,6 @@ import (
 )
 
 // TimeStamp returns the current time in UNIX milliseconds
-var TimeStamp = func() int {
-	return int(time.Now().UnixNano() / 1e6)
+var TimeStamp = func() int64 {
+	return time.Now().UnixNano() / 1e6
 }
